fix(patcher): Match DQXUpdater SqPack version-header + empty-block semantics
Reverse-engineered from DQXUpdater.exe (ZiPatch_WriteSqpackHeader @0x428130,
ZiPatch_SqpkDeleteData_WriteEmptyBlocks @0x427c60) and validated against the
oracle: the boot + game patches 1-4 still reproduce byte-for-byte, and the
SqPack *version header* now matches the updater exactly.

SqpkHeader (onSqpkHeader semantics):
- Index-file header commands are a no-op (the updater only writes Dat headers).
- For the version header (SqPackHeader at offset 0) the updater zeros the
  buildDate/buildTime stamp (8 bytes at 0x18) the patch carries, then recomputes
  the header's SHA-1 self-hash (20 bytes at 0x3C0, over [0,0x3C0)). Replicate both.

SqpackDatFile.WriteEmptyFileBlockAt: write the FileBlockHeader's block count as a
4-byte u32 (XIVLauncher wrote it as an 8-byte long, making the record 24 bytes
instead of the updater's 20).

Still diverging on incremental .dat patches: the updater recomputes each data
header's used-size (0x0C) + SHA-1 (0x20) and truncates partial dats at finalize,
and removes some transient .dat files via its delete list. WIP.

+++ Galapa.Core/Patcher/ZiPatch/Util/SqpackDatFile.cs
@@ -20,16 +20,14 @@ public sealed class SqpackDatFile(BinaryReader reader) : SqpackFile(reader)
 
         using var file = new BinaryWriter(stream, Encoding.Default, true);
 
-        // FileBlockHeader - the 0 writes are technically unnecessary but are in for illustrative purposes
-        // Block size
-        file.Write(1 << 7);
-        // ????
+        // FileBlockHeader: DQXUpdater writes exactly five u32s = [0x80, 0, 0, blockNumber-1, 0]
+        // (the 20-byte record described by ZiPatch_SqpkDeleteData/ExpandData_WriteEmptyBlocks).
+        // NOTE: blockNumber-1 must be a 4-byte int — XIVLauncher writes it as a long, which
+        // would make the record 24 bytes and diverge from the DQX updater.
+        file.Write(1 << 7);                  // block size (0x80)
         file.Write(0);
-        // File size
-        file.Write(0);
-        // Total number of blocks?
-        file.Write(blockNumber - 1);
-        // Used number of blocks?
+        file.Write(0);                       // file size
+        file.Write((int)(blockNumber - 1));  // number of additional empty blocks
         file.Write(0);
     }
 }
